requirements.txt
streamlit
requests
python-dotenv
phidata
ollama
fpdf

